requirements.txt
anthropic>=0.18.0
rich>=13.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
